Fixed static tests
