+++ AdobeIms/Controller/Adminhtml/User/Logout.php
@@ -25,7 +25,7 @@ class Logout extends Action
     /**
      * @see _isAllowed()
      */
-    const ADMIN_RESOURCE = 'Magento_AdobeStockImageAdminUi::save_preview_images';
+    const ADMIN_RESOURCE = 'Magento_AdobeIms::logout';
 
     /**
      * @var CurlFactory
