+++ AdobeIms/Controller/Adminhtml/User/Profile.php
@@ -31,7 +31,7 @@ class Profile extends Action
     /**
      * @see _isAllowed()
      */
-    const ADMIN_RESOURCE = 'Magento_AdobeStockImageAdminUi::save_preview_images';
+    const ADMIN_RESOURCE = 'Magento_AdobeIms::profile';
 
     /**
      * @var UserContextInterface
