+++ AdobeIms/etc/acl.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Acl/etc/acl.xsd">
+    <acl>
+        <resources>
+            <resource id="Magento_Backend::admin">
+                <resource id="Magento_Backend::system">
+                    <resource id="Magento_AdobeIms::adobe_ims" title="Adobe IMS" translate="title" sortOrder="5">
+                        <resource id="Magento_AdobeIms::actions" title="Actions" translate="title">
+                            <resource id="Magento_AdobeIms::login" title="Login" translate="title"/>
+                            <resource id="Magento_AdobeIms::logout" title="Logout" translate="title"/>
+                            <resource id="Magento_AdobeIms::profile" title="Get User Profile" translate="title"/>
+                        </resource>
+                    </resource>
+                </resource>
+            </resource>
+        </resources>
+    </acl>
+</config>
