+++ AdobeStockAsset/composer.json
@@ -6,6 +6,8 @@
         "magento/framework": "*",
         "magento/module-adobe-stock-asset-api": "*",
         "magento/module-adobe-stock-client-api": "*",
+        "magento/module-adobe-media-gallery-api": "*",
+        "magento/module-adobe-media-gallery": "*",
         "magento/module-config": "*"
     },
     "type": "magento2-module",
