+++ AdobeStockAsset/etc/module.xml
@@ -6,5 +6,9 @@
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
-    <module name="Magento_AdobeStockAsset" />
+    <module name="Magento_AdobeStockAsset">
+        <sequence>
+            <module name="Magento_AdobeMediaGallery"/>
+        </sequence>
+    </module>
 </config>
