+++ AdobeStockImage/composer.json
@@ -7,6 +7,7 @@
         "magento/module-adobe-stock-asset-api": "*",
         "magento/module-adobe-stock-client-api": "*",
         "magento/module-adobe-stock-image-api": "*",
+        "magento/module-adobe-media-gallery-api": "*",
         "magento/module-catalog": "*"
     },
     "type": "magento2-module",
