+++ AdobeStockImageAdminUi/Controller/Adminhtml/Preview/RelatedImages.php
@@ -22,7 +22,8 @@ class RelatedImages extends Action
     /**
      * @see _isAllowed()
      */
-    const ADMIN_RESOURCE = 'Magento_Backend::admin';
+    const ADMIN_RESOURCE = 'Magento_AdobeStockImageAdminUi::save_preview_images';
+
     /**
      * Successful get related image result code.
      */
